+++ build-less-to-css.sh
@@ -2,18 +2,10 @@
 
 # Compile CSS
 
-cd themes/clean-base/less
-lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
-cd ../../clean-contrasted/less
-lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
-cd ../../clean-bordered/less
-lessc ./build.less ../css/theme.css --clean-css="--s1 --advanced" --source-map=../css/theme.css.map
-cd ../../..
-
-# Compile dark CSS for the Dark mode module: https://github.com/felixhahnweilheim/humhub-dark-mode
-
 # 1) clean-base (dark)
 cd themes/clean-base
+lessc less/build.less css/theme.css --clean-css="--s1 --advanced" --source-map=css/theme.css.map
+# Compile dark CSS for the Dark mode module: https://github.com/felixhahnweilheim/humhub-dark-mode
 lessc less/dark/build.less css/dark.css
 # Note: Unfortunately the color extractor removes the CSS variables
 css-color-extractor css/dark.css css/dark.css --format=css
@@ -21,10 +13,11 @@ css-color-extractor css/dark.css css/dark.css --format=css
 cp css/dark.css css/temporary.less
 lessc less/dark/build2.less css/dark.css --clean-css="--s1 --advanced"
 rm css/temporary.less
-cd ../../
 
 # 2) clean-bordered (dark)
-cd themes/clean-bordered
+cd ../clean-contrasted
+lessc less/build.less css/theme.css --clean-css="--s1 --advanced" --source-map=css/theme.css.map
+# Compile dark CSS for the Dark mode module: https://github.com/felixhahnweilheim/humhub-dark-mode
 lessc less/dark/build.less css/dark.css
 # Note: Unfortunately the color extractor removes the CSS variables
 css-color-extractor css/dark.css css/dark.css --format=css
@@ -32,10 +25,11 @@ css-color-extractor css/dark.css css/dark.css --format=css
 cp css/dark.css css/temporary.less
 lessc less/dark/build2.less css/dark.css --clean-css="--s1 --advanced"
 rm css/temporary.less
-cd ../../
 
 # 3) clean-contrasted (dark)
-cd themes/clean-contrasted
+cd ../clean-bordered
+lessc less/build.less css/theme.css --clean-css="--s1 --advanced" --source-map=css/theme.css.map
+# Compile dark CSS for the Dark mode module: https://github.com/felixhahnweilheim/humhub-dark-mode
 lessc less/dark/build.less css/dark.css
 # Note: Unfortunately the color extractor removes the CSS variables
 css-color-extractor css/dark.css css/dark.css --format=css
@@ -43,4 +37,5 @@ css-color-extractor css/dark.css css/dark.css --format=css
 cp css/dark.css css/temporary.less
 lessc less/dark/build2.less css/dark.css --clean-css="--s1 --advanced"
 rm css/temporary.less
-cd ../../
+
+cd ../..