+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 --------------------
 - Fix: The top menu "Account" item (at the right) needs right margin on mobile screens
+- Fix: Space of User banner image stretched (change to cover)
 
 1.6.0 (Nov 23, 2023)
 --------------------